Stage 2.4 introduces precise monetary handling by replacing `double` with `BigDecimal` across the application and improving CSV persistence usability.

### Domain
- `Transaction.amount` migrated from `double` to `BigDecimal`
- Monetary values normalized to scale 2 (`RoundingMode.HALF_UP`)

### Query / Filter
- Amount range filtering (`minAmount`, `maxAmount`) migrated to `BigDecimal`
- Comparisons updated to use `BigDecimal` semantics

### Reports
- Report DTOs and services updated to use `BigDecimal`
- Monetary output consistently formatted to two decimal places via `ConsolePrinter`

### CSV persistence
- CSV parsing now reads amounts as `BigDecimal`
- CSV output stores monetary values as plain decimal text
- Verified round-trip correctness

### CLI improvements
- Save and Load commands support custom filename or path
- Default fallback filename: `transactions.csv`
- Missing directories are created automatically on save

### Tests
- Unit tests updated to work with `BigDecimal`
- Added money correctness test (`0.10 + 0.20 = 0.30`)
- CSV persistence tests updated

### Documentation
- Added manual test cases for Stage 2.4:
  - `docs/manual-tests/manual-tests-stage-2.4.md`
  - `docs/manual-tests/manual-tests-stage-2.4.csv`
- README updated with BigDecimal migration details

+++ src/main/java/com/owsiankagrzegorz/expensetracker/app/ExpenseTrackerApp.java
@@ -10,6 +10,8 @@ import com.owsiankagrzegorz.expensetracker.repository.InMemoryTransactionReposit
 import com.owsiankagrzegorz.expensetracker.service.ExpenseTrackerService;
 import com.owsiankagrzegorz.expensetracker.service.query.*;
 import com.owsiankagrzegorz.expensetracker.service.report.TransactionReportService;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Scanner;
 
@@ -25,9 +27,9 @@ public class ExpenseTrackerApp {
 
         boolean seed = Boolean.getBoolean("seed");
         if (seed) {
-            service.addTransaction(new Transaction(1L, 100.0, "Jedzenie", LocalDate.of(2026, 1, 1), TransactionType.WYDATEK));
-            service.addTransaction(new Transaction(2L, 200.0, "Transport", LocalDate.of(2026, 1, 1), TransactionType.WYDATEK));
-            service.addTransaction(new Transaction(3L, 300.0, "Przychód", LocalDate.of(2026, 1, 1), TransactionType.PRZYCHOD));
+            service.addTransaction(new Transaction(1L, BigDecimal.valueOf(100.0), "Jedzenie", LocalDate.of(2026, 1, 1), TransactionType.WYDATEK));
+            service.addTransaction(new Transaction(2L, BigDecimal.valueOf(200.0), "Transport", LocalDate.of(2026, 1, 1), TransactionType.WYDATEK));
+            service.addTransaction(new Transaction(3L, BigDecimal.valueOf(300.0), "Przychód", LocalDate.of(2026, 1, 1), TransactionType.PRZYCHOD));
         }
 
         InputReader input = new InputReader(new Scanner(System.in));
